test1/pi/requirements.txt
picamera2
requests
Pillow